fix(EmailDialog): fixes dialog not closing on browser back
